+++ techgrains/lib/com/techgrains/listener/tg_session_listener.dart
@@ -1,0 +1,11 @@
+/// Listener of TGSession
+abstract class TGSessionListener {
+  /// When TGSession invalidated
+  void invalidate();
+
+  /// When put key in TGSession
+  void keySet(String key);
+
+  /// When key removed from TGSession
+  void keyRemove(String key);
+}